deleting zip file

+++ configure-gateway.sh
@@ -26,8 +26,13 @@ if [ ! -f $DOWNLOAD_ZIP_FILE_NAME ]
 	    sed -i 's/PASSWORD=""/PASSWORD="${WSO2_CLOUD_PASSWORD}"/' $UNZIPPED_FILE_NAME/bin/configure-gateway.sh
 	    echo 'sed -i "s/8280/$PORT/" wso2am-2.1.0/repository/conf/axis2/axis2.xml' >> $UNZIPPED_FILE_NAME/bin/configure-gateway.sh
 	    echo 'bash wso2am-2.1.0/bin/wso2server.sh' >> $UNZIPPED_FILE_NAME/bin/configure-gateway.sh
-            rm -rf DOWNLOAD_ZIP_FILE_NAME
         fi
 fi
-ls
 sh $UNZIPPED_FILE_NAME/bin/configure-gateway.sh
+if [  -f $DOWNLOAD_ZIP_FILE_NAME ]; then
+    echo "###################deleting####################"
+    rm -rf DOWNLOAD_ZIP_FILE_NAME
+    
+fi
+
+ls-lh
